ref: make `NewClient()` api wrapper accept token as input
It feels like the proper thing to do is have the NewClient() wrapper
func accept the token as a struct field param to use to configure the
api client similar to the other params, rather than have this internal
wrapper func find and retrieve the token itself from the env.

This refactors things to surface the token retreival to the main package
in the `Run()` func for use with creating the api client to provide to
the exporter/sd workers.